🛰️ Enhanced organization of SSE API
